--- BloodFlowTrixi.jl/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Redirecting to BloodFlowTrixi.jl documentation...</title>
-  <link rel="canonical" href="https://juliahealth.github.io/BloodFlowTrixi.jl/stable/">
-  <meta http-equiv="refresh" content="0; url=https://juliahealth.github.io/BloodFlowTrixi.jl/stable/">
-  <script type="text/javascript">
-    window.location.href = "https://juliahealth.github.io/BloodFlowTrixi.jl/stable/";
-  </script>
-</head>
-<body>
-  <p>If you are not redirected automatically, follow this
-    <a href="https://juliahealth.github.io/BloodFlowTrixi.jl/stable/">link to BloodFlowTrixi.jl documentation</a>.
-  </p>
-</body>
-</html>
