--- MedImages.jl/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Redirecting to MedImages.jl documentation...</title>
-  <link rel="canonical" href="https://juliahealth.github.io/MedImages.jl/dev/">
-  <meta http-equiv="refresh" content="0; url=https://juliahealth.github.io/MedImages.jl/dev/">
-  <script type="text/javascript">
-    window.location.href = "https://juliahealth.github.io/MedImages.jl/dev/";
-  </script>
-</head>
-<body>
-  <p>If you are not redirected automatically, follow this
-    <a href="https://juliahealth.github.io/MedImages.jl/dev/">link to MedImages.jl documentation</a>.
-  </p>
-</body>
-</html>
