--- docs/DICOMTree.jl/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Redirecting to DICOMTree.jl documentation...</title>
-  <link rel="canonical" href="https://juliapackages.com/p/dicomtree/">
-  <meta http-equiv="refresh" content="0; url=https://juliapackages.com/p/dicomtree/">
-  <script type="text/javascript">
-    window.location.href = "https://juliapackages.com/p/dicomtree/";
-  </script>
-</head>
-<body>
-  <p>If you are not redirected automatically, follow this
-    <a href="https://juliapackages.com/p/dicomtree/">link to DICOMTree.jl documentation</a>.
-  </p>
-</body>
-</html>
