--- docs/NeuroAnalyzer.jl/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Redirecting to NeuroAnalyzer.jl documentation...</title>
-  <link rel="canonical" href="https://neuroanalyzer.org/">
-  <meta http-equiv="refresh" content="0; url=https://neuroanalyzer.org/">
-  <script type="text/javascript">
-    window.location.href = "https://neuroanalyzer.org/";
-  </script>
-</head>
-<body>
-  <p>If you are not redirected automatically, follow this
-    <a href="https://neuroanalyzer.org/">link to NeuroAnalyzer.jl documentation</a>.
-  </p>
-</body>
-</html>
